Added basic Spansh support

+++ EliteAPI.Events/Status/Ship/Events/HeadingStatusEvent.cs
@@ -2,11 +2,11 @@
 
 namespace EliteAPI.Events.Status.Ship.Events;
 
-public readonly struct HeadingStatusEvent : IStatusEvent<float>
+public readonly struct HeadingStatusEvent : IStatusEvent<double>
 {
     public DateTime Timestamp => DateTime.Now;
     
     public string Event => "Heading";
     
-    public float Value { get; init; }
+    public double Value { get; init; }
 }
