+++ EliteAPI.Events/Status/Ship/Events/HealthStatusEvent.cs
@@ -2,11 +2,11 @@
 
 namespace EliteAPI.Events.Status.Ship.Events;
 
-public readonly struct HealthStatusEvent : IStatusEvent<float>
+public readonly struct HealthStatusEvent : IStatusEvent<double>
 {
     public DateTime Timestamp => DateTime.Now;
     
     public string Event => "Health";
     
-    public float Value { get; init; }
+    public double Value { get; init; }
 }
