+++ EliteAPI.Events/Status/Ship/Events/LatitudeStatusEvent.cs
@@ -2,11 +2,11 @@
 
 namespace EliteAPI.Events.Status.Ship.Events;
 
-public readonly struct LatitudeStatusEvent : IStatusEvent<float>
+public readonly struct LatitudeStatusEvent : IStatusEvent<double>
 {
     public DateTime Timestamp => DateTime.Now;
     
     public string Event => "Latitude";
     
-    public float Value { get; init; }
+    public double Value { get; init; }
 }
