+++ EliteAPI.Events/Status/Ship/Events/TemperatureStatusEvent.cs
@@ -2,11 +2,11 @@
 
 namespace EliteAPI.Events.Status.Ship.Events;
 
-public readonly struct TemperatureStatusEvent : IStatusEvent<float>
+public readonly struct TemperatureStatusEvent : IStatusEvent<double>
 {
     public DateTime Timestamp => DateTime.Now;
     
     public string Event => "Temperature";
     
-    public float Value { get; init; }
+    public double Value { get; init; }
 }
